cleaned up some routes

+++ app/api/routes2.py
@@ -1,5 +1,6 @@
 from flask import Flask,jsonify,request,url_for,render_template,session,current_app
 import requests
+from flask_login import current_user
 from notes01.app import login_manager,db,getuploadpath
 from . import api,myapi
 import requests,sys
@@ -15,16 +16,24 @@ def find_course(name):
     course=Courses.query.filter_by(name=name).first()
     return course
 
-def find_unit(unit_prefix):
-        unit=Units.query.filter_by(acronym=unit_prefix).first()
+def find_unit(code):
+        unit=Units.query.filter_by(acronym=code).first()
         if unit:
-            return (unit.id)
+            return (unit)
         else:
             return (None)     
         
+def find_file(name):
+    exists=Notes.query.filter_by(name=name).first()
+    if exists:
+        return True
+    else:
+        return False
 
-              
-def savefile(file,unit_prefix='',toupload=False):
+
+    
+# reason for passing code as a param even though its not used is for local server storage if its ever implemented
+def savefile(file,code='',toupload=False):
     if toupload:
         path=getuploadpath()
         try:
@@ -44,37 +53,72 @@ def get_notes(course_name):
             for n in unit.notes:
                 notes.append({'unit':unit.acronym,'name':n.name,'gid':n.gid})
         if len(notes)>0:
-            return jsonify({course_name:notes})
+            return notes
     else:
         return None
-    return None#runs only if the notes list is zero
 
 
  
 
 
+@api.route('/home')
+@api.route('/')
+def home():
+    if current_user.is_authenticated:
+        return render_template('landing.html',title='Biblioteca')
+    else:
+        return render_template('documentation.html',title='Biblioteca')
+
+                           
 @api.route('/new_course',methods=['GET'])
 def new_course():
-    return render_template('addcourse.html')
+    return render_template('addcourse.html',title='Addcourse')
 
 
 @api.route('/find_notes',methods=['GET'])
 def find_notes():
-    return render_template('notes.html')
+    return render_template('notes.html',title='Notes')
     
 
 @api.route('/upload_notes',methods=['GET'])
 def upload_notes():
-    return render_template('upload.html')
+    return render_template('upload.html',title='Uploader')
+# @api.route('/admin',methods=['GET','POST'])
+# def adminpanel():
+#     if current_user.is_authenticated and current_user.is_admin:
+#         return render_template('admin/admin.html')
+#     else:
+#         return redirect(url_for('auth.home'))
 
 parser=reqparse.RequestParser()
 parser.add_argument('course_name',type=str,help='course name required')
 
-acmodel=myapi.model('AddCourse',{'name':fields.String()})
+unitparser=reqparse.RequestParser()
+unitparser.add_argument('unit',type=str,help='unit acronym required')
+
+acmodel=myapi.model('AddCourse',{'course_name':fields.String()})
 cdmodel=myapi.model('CourseDetails',{'email':fields.String()})
 amcmodel=myapi.model('AddMyCourse',{'email':fields.String(),'course_name':fields.String()})
 cnmodel=myapi.model('CourseNotes',{'course_name':fields.String()})
+unmodel=myapi.model('UnitNotes',{'unit_code':fields.String()})
+# uploadmodel=myapi.model('Upload',{'headers':'unit'})
 
+class UnitNotes(Resource):
+    @myapi.expect(unmodel)
+    @myapi.doc(body=unmodel)
+    def post(self):
+        data=unitparser.parse_args()
+        print(data)
+        if(code:=data.get('unit_code')):
+            unit=find_unit(code)
+            if unit:
+                return {'notes':[note.name for note in unit.notes]}
+            else:
+                return {'error':'unit not found'}
+        else:
+            return {'error':"invalid information recieved,expected json obj unit:unit acronym"}
+            
+    
 class CourseNotes(Resource):
     @myapi.expect(cnmodel)
     @myapi.doc(body=cnmodel)
@@ -95,11 +139,10 @@ class AddCourse(Resource):
     @myapi.expect(acmodel)
     def post(self):
         data=request.json
-        if data.get('course_name'):
-            course_name=data.get('course_name')
+        if (course_name:=data.get('course_name')):
             course=find_course(course_name)
             if course:
-                return {'error':'course already exists'}
+                return {'message':'course already exists'}
             else:
                 c=Courses(name='course_name')
                 db.session.add(c)
@@ -109,13 +152,13 @@ class AddCourse(Resource):
                 except:
                     db.session.rollback()
         else:
-            return {'error':'invalid course information'}
+            return {'error':'invalid course information in request expected json obj {course name:string}'}
 
 class Uploadlocal(Resource):
     def post(self):
         if 'files' not in request.files:
             return jsonify({'error':' no file was chosen'})
-        acronym=request.form.get('unit_prefix')
+        acronym=request.form.get('unit_code')
         for file in request.files.getlist('files'):
             res=savefile(file,acronym)
             if not res:
@@ -125,36 +168,42 @@ class Uploadlocal(Resource):
         return jsonify({'success':'request complete'})  
      
 class Upload(Resource):
+    
     def post(self):
        
         files=request.files.getlist('notes')
-        unit=request.headers.get('unit') or request.form.get('unit_prefix')
-        unit_id=find_unit(unit)
+        code=request.headers.get('unit_code') or request.form.get('unit_code')
+        unit_id=find_unit(code).id
+        print(unit_id)
         
         if files:
             creds=FileUploader.getcreds()
             for file in files:
-                res=savefile(file,unit,toupload=True)
+                res=savefile(file,code,toupload=True)
                 if res.get('error'):
                     return res,500
                 if (path:=res.get('path')):
-                    obj=FileUploader(path,file.filename,unit)
-                    id=obj.driveupload(creds)
+                    obj=FileUploader(path,file.filename,code)
+                    exists=find_file(obj.name)
+                    if exists:
+                        continue
+                    obj.driveupload(creds)
+                    print(obj.id)
                     obj.delete_file()
-                    # res=obj.getResource()
-                    if id:
-                       note=Notes(name=obj.name,gid=str(obj.id),unit_id=unit_id)
+                    if unit_id and obj.id:
+                       note=Notes(name=obj.name,gid=obj.id,unit_id=unit_id)
                        db.session.add(note)
                        try:
                            db.session.commit()
-                           return {'success':'file uploaded succesfully'}
                        except Exception as e:
+                           print(e)
                            db.session.rollback()
                            return {'error':sys.exc_info()[0]}
                     else:
                         return {'error':'an error occured try again later'}
                     
-        return
+        return {'success':'file(s) uploaded succesfully'}
+
       
 class AddMyCourse(Resource):
     @myapi.expect(amcmodel)
@@ -217,6 +266,12 @@ class AllCourses(Resource):
         courses=Courses.query.all()
         data=[course.name for course in courses]
         return {"courses":data}
+    
+class AllUnits(Resource):
+    def get(self):
+        units=Units.query.all()
+        data=[{'code':unit.acronym,'name':unit.name,'year':unit.year,'semester':unit.semester} for unit in units]
+        return data
 class AllNotes(Resource):
     def get(self):
         notes=Notes.query.all()
@@ -229,7 +284,7 @@ class GetUnits(Resource):
         if (name:=data.get('course_name')):
             course=find_course(name)
             if course:
-                return {course.name:[unit.acronym for unit in course.units]}
+                return [{'acronym':unit.acronym,'name':unit.name,'year':unit.year,'semester':unit.semester} for unit in course.units]
             else:
                 return {'error':'course unavailable'}
         else:
@@ -240,9 +295,12 @@ myapi.add_resource(AllCourses,'/courses')
 myapi.add_resource(CourseNotes,'/course_notes')
 myapi.add_resource(AddCourse,'/add_course')
 myapi.add_resource(CourseDetails,'/course_details')
-myapi.add_resource(GetUnits,'/units')
-myapi.add_resource(Upload,'/upload_notes')
+myapi.add_resource(GetUnits,'/course_units')
+myapi.add_resource(Upload,'/upload')
 myapi.add_resource(AddMyCourse,'/add_mycourse')
 myapi.add_resource(AllNotes,'/all_notes')
+myapi.add_resource(UnitNotes,'/unit_notes')
+myapi.add_resource(AllUnits,'/all_units')
+
